requirements.txt
psycopg2==2.9.7
python-slugify==8.0.1
python-dotenv==1.0.0
prefect==2.13.0
